+++ src/test/scala/com/github/swagger/akka/samples/DictHttpService.scala
@@ -1,4 +1,4 @@
-package com.github.swagger.spray.samples
+package com.github.swagger.akka.samples
 
 import io.swagger.annotations._
 import javax.ws.rs.Path
